+++ helix-core/src/main/java/org/apache/helix/monitoring/mbeans/MessageLatencyMonitor.java
@@ -32,6 +32,7 @@ import org.apache.helix.monitoring.mbeans.dynamicMBeans.DynamicMetric;
 import org.apache.helix.monitoring.mbeans.dynamicMBeans.HistogramDynamicMetric;
 import org.apache.helix.monitoring.mbeans.dynamicMBeans.SimpleDynamicMetric;
 
+
 public class MessageLatencyMonitor extends DynamicMBeanProvider {
   private static final String MBEAN_DESCRIPTION = "Helix Message Latency Monitor";
   private final String _sensorName;
