+++ helix-core/src/main/java/org/apache/helix/monitoring/mbeans/RoutingTableProviderMonitor.java
@@ -34,6 +34,7 @@ import org.apache.helix.monitoring.mbeans.dynamicMBeans.DynamicMetric;
 import org.apache.helix.monitoring.mbeans.dynamicMBeans.HistogramDynamicMetric;
 import org.apache.helix.monitoring.mbeans.dynamicMBeans.SimpleDynamicMetric;
 
+
 public class RoutingTableProviderMonitor extends DynamicMBeanProvider {
   public static final String DATA_TYPE_KEY = "DataType";
   public static final String CLUSTER_KEY = "Cluster";
